test/diskfont: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@51786 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/diskfont/fontinfo.c
@@ -159,7 +159,7 @@ static void openfont(void)
 	printf("  taglist:\n");
 	while((tag = NextTagItem(&tstate)))
 	{
-            printf("    {%08x,%p}\n", tag->ti_Tag, (void *)tag->ti_Data);
+            printf("    {%08x,%p}\n", (unsigned)tag->ti_Tag, (void *)tag->ti_Data);
 	}
 	
 	dpivalue = GetTagData(TA_DeviceDPI, 0, extension);
